requirements.txt
faunadb==4.1.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
gunicorn==20.1.0
python-dotenv==0.19.0
pytz==2021.1
